Cleaning cache needs to be forced
As of npm@5, the npm cache self-heals from corruption issues and data extracted from the cache is guaranteed to be valid. If you want to make sure everything is consistent, use 'npm cache verify' instead.

If you're sure you want to delete the entire cache, rerun this command with --force.